requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=1_gnu
absl-py=0.13.0=pyhd3eb1b0_0
aiohttp=3.7.4=py37h27cfd23_1
albumentations=1.0.3=pyhd8ed1ab_0
async-timeout=3.0.1=py37h06a4308_0
attrs=21.2.0=pyhd3eb1b0_0
blas=1.0=mkl
blinker=1.4=py37h06a4308_0
bottleneck=1.3.2=py37heb32a55_1
brotlipy=0.7.0=py37h27cfd23_1003
bzip2=1.0.8=h7f98852_4
c-ares=1.17.1=h27cfd23_0
ca-certificates=2021.7.5=h06a4308_1
cachetools=4.2.2=pyhd3eb1b0_0
cairo=1.16.0=hcf35c78_1003
certifi=2021.5.30=py37h06a4308_0
cffi=1.14.0=py37h2e261b9_0
chardet=3.0.4=py37h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.1=pyhd3eb1b0_0
cloudpickle=1.6.0=py_0
coverage=5.5=py37h27cfd23_2
cryptography=3.4.7=py37hd23ed53_0
cudatoolkit=10.0.130=0
cycler=0.10.0=py37_0
cython=0.29.24=py37h295c915_0
cytoolz=0.11.0=py37h7b6447c_0
dask-core=2021.8.0=pyhd3eb1b0_0
dbus=1.13.6=he372182_0
expat=2.4.1=h9c3ff4c_0
ffmpeg=4.3.2=hca11adc_0
fontconfig=2.13.1=h6c09931_0
freeglut=2.8.1=0
freetype=2.10.4=h5ab3b9f_0
fsspec=2021.7.0=pyhd3eb1b0_0
geos=3.8.0=he6710b0_0
giflib=5.2.1=h7b6447c_0
glib=2.63.1=h5a9c865_0
gmp=6.2.1=h58526e2_0
gnutls=3.6.13=h85f3911_1
google-auth=1.33.0=pyhd3eb1b0_0
google-auth-oauthlib=0.4.1=py_2
graphite2=1.3.14=h23475e2_0
grpcio=1.36.1=py37h2157cd5_1
gst-plugins-base=1.14.5=h0935bb2_2
gstreamer=1.14.5=h36ae1b5_2
harfbuzz=2.4.0=h9f30f68_3
hdf5=1.10.6=nompi_h3c11f04_101
icu=64.2=he1b5a44_1
idna=3.2=pyhd3eb1b0_0
imageio=2.9.0=pyhd3eb1b0_0
imgaug=0.4.0=pyhd3eb1b0_0
importlib-metadata=3.10.0=py37h06a4308_0
intel-openmp=2021.3.0=h06a4308_3350
jasper=1.900.1=h07fcdf6_1006
joblib=1.0.1=pyhd8ed1ab_0
jpeg=9d=h36c2ea0_0
kiwisolver=1.3.1=py37h2531618_0
lame=3.100=h7f98852_1001
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.35.1=h7274673_9
libblas=3.9.0=11_linux64_openblas
libcblas=3.9.0=11_linux64_openblas
libclang=9.0.1=cling_v0.9_hc213394_1
libedit=3.1.20210714=h7f8727e_0
libffi=3.2.1=1
libgcc-ng=11.2.0=h1d223b6_8
libgfortran-ng=7.3.0=hdf63c60_0
libgfortran5=11.2.0=h5c6108e_8
libglu=9.0.0=hf484d3e_1
libgomp=11.2.0=h1d223b6_8
libiconv=1.16=h516909a_0
liblapack=3.9.0=11_linux64_openblas
liblapacke=3.9.0=11_linux64_openblas
libllvm9=9.0.1=default_hc23dcda_4
libopenblas=0.3.17=pthreads_h8fe5266_1
libopencv=4.5.0=py37_3
libopus=1.3.1=h7b6447c_0
libpng=1.6.37=hbc83047_0
libprotobuf=3.17.2=h4ff587b_1
libstdcxx-ng=11.2.0=he4da1e4_8
libtiff=4.2.0=h85742a9_0
libuuid=1.0.3=0
libvpx=1.7.0=h439df22_0
libwebp=1.2.0=h89dd481_0
libwebp-base=1.2.0=h27cfd23_0
libxcb=1.14=h7b6447c_0
libxkbcommon=0.10.0=he1b5a44_0
libxml2=2.9.10=hee79883_0
locket=0.2.1=py37h06a4308_1
lz4-c=1.9.3=h295c915_1
markdown=3.3.4=py37h06a4308_0
matplotlib-base=3.3.4=py37h62a2d02_0
mkl=2021.3.0=h06a4308_520
mkl-service=2.4.0=py37h7f8727e_0
mkl_fft=1.3.0=py37h42c9631_2
mkl_random=1.2.2=py37h51133e4_0
multidict=5.1.0=py37h27cfd23_2
ncurses=6.2=he6710b0_1
nettle=3.6=he412f7d_0
networkx=2.6.2=pyhd3eb1b0_0
ninja=1.7.2=0
nspr=4.30=h9c3ff4c_0
nss=3.59=h2c00c37_0
numexpr=2.7.3=py37h22e1b3c_1
numpy=1.20.3=py37hf144106_0
numpy-base=1.20.3=py37h74d4b33_0
oauthlib=3.1.1=pyhd3eb1b0_0
olefile=0.46=py_0
opencv=4.5.0=py37_3
openh264=2.1.1=h780b84a_0
openssl=1.1.1k=h27cfd23_0
packaging=21.0=pyhd3eb1b0_0
pandas=1.3.2=py37h8c16a72_0
partd=1.2.0=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pillow=8.3.1=py37h5aabda8_0
pip=21.2.2=py37h06a4308_0
pixman=0.38.0=h516909a_1003
protobuf=3.17.2=py37h295c915_0
py-opencv=4.5.0=py37h888b3d9_3
pyasn1=0.4.8=py_0
pyasn1-modules=0.2.8=py_0
pycparser=2.20=py_2
pyjwt=2.1.0=py37h06a4308_0
pyopenssl=20.0.1=pyhd3eb1b0_1
pyparsing=2.4.7=pyhd3eb1b0_0
pysocks=1.7.1=py37_1
python=3.7.7=h191fe78_0_cpython
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.7=2_cp37m
pytorch=1.2.0=py3.7_cuda10.0.130_cudnn7.6.2_0
pytz=2021.1=pyhd3eb1b0_0
pywavelets=1.1.1=py37h7b6447c_2
pyyaml=5.4.1=py37h27cfd23_1
qt=5.12.5=hd8c4c69_1
readline=7.0=h7b6447c_5
requests=2.26.0=pyhd3eb1b0_0
requests-oauthlib=1.3.0=py_0
rsa=4.7.2=pyhd3eb1b0_1
scikit-image=0.18.1=py37ha9443f7_0
scikit-learn=0.24.2=py37hf0f1638_1
scipy=1.6.2=py37had2a1c9_1
setuptools=52.0.0=py37h06a4308_0
shapely=1.7.1=py37h98ec03d_0
six=1.16.0=pyhd3eb1b0_0
sqlite=3.33.0=h62c20be_0
tensorboard=2.5.0=py_0
tensorboard-plugin-wit=1.6.0=py_0
threadpoolctl=2.2.0=pyh8a188c0_0
tifffile=2020.10.1=py37hdd07704_2
tk=8.6.10=hbc83047_0
toolz=0.11.1=pyhd3eb1b0_0
torchvision=0.4.0=py37_cu100
tornado=6.1=py37h27cfd23_0
typing-extensions=3.10.0.0=hd3eb1b0_0
typing_extensions=3.10.0.0=pyh06a4308_0
urllib3=1.26.6=pyhd3eb1b0_1
werkzeug=1.0.1=pyhd3eb1b0_0
wheel=0.37.0=pyhd3eb1b0_0
x264=1!161.3030=h7f98852_1
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.0.10=h7f98852_0
xorg-libsm=1.2.2=h470a237_5
xorg-libx11=1.7.2=h7f98852_0
xorg-libxext=1.3.4=h7f98852_1
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h7f98852_1002
xorg-xproto=7.0.31=h7f98852_1007
xz=5.2.5=h7b6447c_0
yaml=0.2.5=h7b6447c_0
yarl=1.6.3=py37h27cfd23_0
zipp=3.5.0=pyhd3eb1b0_0
zlib=1.2.11=0
zstd=1.4.9=haebb681_0
